requirements.txt
pandas
pyyaml
google-generativeai
